Extract patch to function
